Rename sentence counting function file and test further.
