+++ services/QuillComprehension/app/javascript/src/lib/countSentences.ts
@@ -1,0 +1,6 @@
+import nlp from 'compromise';
+
+export function countSentences(text:string):number{
+  const doc = nlp(text, null);
+  return doc.sentences().length
+}
